t/16-coverage.t: expand tests for binsearch_range()
This perl library is ancient and not maintained, it also has a license
which is not compatible with the GPL. The code is very convoluted, so
to make it compatible with super ancient versions of perl. So let's just
help people who use it replace it in their own projects with a set of
tests, so to provide a replacement GPL compatible set of helpers.

This just expands on binsearch_range() as that is the use case we are
interested in.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>